Add files via upload
These MATLAB functions can be used in functional connectivity analysis through ICA. 

RAICAR (Ranking and averaging independent component analysis by reproducibility) is a method to order components based on their reproducibility in several runs. This function is modified of the main version, described in Z.Yang, et al. 2008. 

The edge pruning algorithm defines the important connections of an components based on their effects on ICA reversibility. The most important edges are kept and the others are replaced by zero.